Alot of Flag shit
Testing with claimers. getting some logic sorted. At the moment if a flag called "HarvestSources" is placed it will init and send a harvester into the room. It was calling init source but I was getting container errors D: so diablesssss it. The visitor function looping in rooms picks up the room.
Also when needs met you get a energy out of energyCapacity in the debug log.