some fixes and code clean up
Fixed the filter function up.
added the code so the builder can get another task.
fiddled with constantTasks for harvesting and container mining. Now will only spawn containerMiner when there is a container , Same goes for miner task declairation. Only when there is a container next to a source..
room.prototype.avliableCreeps removed as is now more a memory based thang.