The start of somthing big. Just gotta lay it out ;)
So here we go. Added to Index.js for Layout
an init that just organises some ground work for memory
same with a build list

changed the range on the dirtroads to not place ontop of shit

created a buildque that cycles through the default template and goes for gold. Atm it places them all at once. Because yanno, I just want it all. Right now. 3 days of uh and oo.

this file will be restructured in the near future to be laid out and have coments and whatnot